+++ app/src/main/java/com/robertconstantindinescu/woutapp/core/presentation/MainActivity.kt
@@ -31,8 +31,8 @@ import com.robertconstantindinescu.woutapp.core.presentation.navigation.screen.B
 import com.robertconstantindinescu.woutapp.core.presentation.ui.theme.WoutAppTheme
 import com.robertconstantindinescu.woutapp.feature_authentication.presentation.register.SignUpScreen
 import com.robertconstantindinescu.woutapp.feature_create_post.presentation.CreatePostScreen
-import com.robertconstantindinescu.woutapp.feature_get_posts.presentation.main_feed_screen.MainFeedScreen
-import com.robertconstantindinescu.woutapp.feature_get_posts.presentation.personal_screen.PersonalScreen
+import com.robertconstantindinescu.woutapp.feature_posts.presentation.main_feed_screen.MainFeedScreen
+import com.robertconstantindinescu.woutapp.feature_posts.presentation.personal_screen.PersonalScreen
 import com.robertconstantindinescu.woutapp.feature_splash.presentation.SplashScreen
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
@@ -43,8 +43,8 @@ class MainActivity : ComponentActivity() {
     @Inject
     lateinit var imageLoader: ImageLoader
 
-    @ExperimentalMaterial3Api
-    @OptIn(ExperimentalPagerApi::class)
+    @OptIn(ExperimentalPagerApi::class, ExperimentalMaterial3Api::class)
+    //@OptIn(ExperimentalPagerApi::class)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
