--- app/src/main/java/com/robertconstantindinescu/woutapp/feature_get_posts/domain/use_case/PersonalGetAllPostsUseCase.kt
@@ -1,12 +1,0 @@
-package com.robertconstantindinescu.woutapp.feature_get_posts.domain.use_case
-
-import com.robertconstantindinescu.woutapp.core.util.Resource
-import com.robertconstantindinescu.woutapp.feature_get_posts.domain.model.PostDM
-import com.robertconstantindinescu.woutapp.feature_get_posts.domain.repository.PostRepository
-import com.robertconstantindinescu.woutapp.feature_get_posts.util.MainFeedConstants.DEFAULT_PAGE_SIZE
-
-class PersonalGetAllPostsUseCase(private val repository: PostRepository) {
-
-    suspend operator fun invoke(page: Int, pageSize: Int = DEFAULT_PAGE_SIZE):
-            Resource<List<PostDM>> = repository.getAllCurrentUserPosts(page, pageSize)
-}
