--- app/src/main/java/com/robertconstantindinescu/woutapp/feature_get_posts/domain/use_case/PostUseCases.kt
@@ -1,6 +1,0 @@
-package com.robertconstantindinescu.woutapp.feature_get_posts.domain.use_case
-
-data class PostUseCases(
-    val mainFeedGetAllPostsUseCase: MainFeedGetAllPostsUseCase,
-    val personalGetAllPostsUseCase: PersonalGetAllPostsUseCase
-)
