--- app/src/main/java/com/robertconstantindinescu/woutapp/feature_get_posts/util/MainFeedConstants.kt
@@ -1,5 +1,0 @@
-package com.robertconstantindinescu.woutapp.feature_get_posts.util
-
-object MainFeedConstants {
-    const val DEFAULT_PAGE_SIZE = 15
-}
