+++ app/src/main/java/com/robertconstantindinescu/woutapp/feature_posts/data/util/Constants.kt
@@ -1,0 +1,7 @@
+package com.robertconstantindinescu.woutapp.feature_posts.data.util
+
+object Constants {
+
+    const val FAVORITE_POST_TABLE = "favorite_post_table"
+    const val POST_FAVORITE_DATABASE_NAME = "favorite_post_db"
+}
