+++ app/src/main/java/com/robertconstantindinescu/woutapp/feature_posts/domain/use_case/MainFeedGetAllPostsUseCase.kt
@@ -1,0 +1,12 @@
+package com.robertconstantindinescu.woutapp.feature_posts.domain.use_case
+
+import com.robertconstantindinescu.woutapp.core.util.Resource
+import com.robertconstantindinescu.woutapp.feature_posts.domain.model.PostDM
+import com.robertconstantindinescu.woutapp.feature_posts.domain.repository.PostRepository
+import com.robertconstantindinescu.woutapp.feature_posts.util.MainFeedConstants.DEFAULT_PAGE_SIZE
+
+class MainFeedGetAllPostsUseCase(private val repository: PostRepository) {
+
+    suspend operator fun invoke(page: Int, pageSize: Int = DEFAULT_PAGE_SIZE):
+            Resource<List<PostDM>> = repository.getAllPosts(page, pageSize)
+}
