+++ app/src/main/java/com/robertconstantindinescu/woutapp/feature_posts/domain/use_case/PostUseCases.kt
@@ -1,0 +1,9 @@
+package com.robertconstantindinescu.woutapp.feature_posts.domain.use_case
+
+data class PostUseCases(
+    val mainFeedGetAllPostsUseCase: MainFeedGetAllPostsUseCase,
+    val personalGetAllPostsUseCase: PersonalGetAllPostsUseCase,
+    val insertPostIntoFavoritesUseCase: InsertPostIntoFavoritesUseCase,
+    val deleteFromFavoritesUseCase: DeleteFromFavoritesUseCase,
+    val getAllFavoritesPostsUseCase: GetAllFavoritesPostsUseCase
+)
