+++ app/src/main/java/com/robertconstantindinescu/woutapp/feature_posts/presentation/common/mapper/PostPresentationMapper.kt
@@ -1,0 +1,35 @@
+package com.robertconstantindinescu.woutapp.feature_posts.presentation.common.mapper
+
+import com.robertconstantindinescu.woutapp.feature_posts.domain.model.PostDM
+import com.robertconstantindinescu.woutapp.feature_posts.presentation.main_feed_screen.model.PostVO
+
+fun PostDM.toPostVO(): PostVO =
+    PostVO(
+        postId = postId,
+        userId = userId,
+        userName = userName,
+        imageUrl = imageUrl,
+        profileImage = profileImage,
+        sportType = sportType,
+        description = description,
+        location = location,
+        subscriptionsCount = subscriptionsCount,
+        isAddedToFavorites = isAddedToFavorites,
+        isUserSubscribed = isUserSubscribed
+    )
+
+
+fun PostVO.toPostDM(): PostDM =
+    PostDM(
+        postId = postId,
+        userId = userId,
+        userName = userName,
+        imageUrl = imageUrl,
+        profileImage = profileImage,
+        sportType = sportType,
+        description = description,
+        location = location,
+        subscriptionsCount = subscriptionsCount,
+        isAddedToFavorites = isAddedToFavorites,
+        isUserSubscribed = isUserSubscribed
+    )
