+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
     ext {
-        compose_version = '1.0.1'
+        compose_version = '1.2.0-alpha01'
         accompanist_version = '0.21.2-beta'
     }
     repositories {
@@ -11,7 +11,7 @@ buildscript {
     }
     dependencies {
         classpath "com.android.tools.build:gradle:7.0.4"
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.21"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.10"
         classpath "com.google.dagger:hilt-android-gradle-plugin:2.40"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
